+++ lib/compony.rb
@@ -96,6 +96,7 @@ module Compony
 
   # Generates a Rails path to a component. Examples: `Compony.path(:index, :users)`, `Compony.path(:show, User.first)`
   # @param comp_name_or_cst [String,Symbol] The component that should be loaded, for instance `ShowForAll`, `'ShowForAll'` or `:show_for_all`
+  #                         or can also pass a component class (such as Components::Users::Show)
   # @param model_or_family_name_or_cst [String,Symbol,ApplicationRecord] Either the family that contains the requested component,
   #                                    or an instance implementing `model_name` from which the family name is auto-generated. Examples:
   #                                    `Users`, `'Users'`, `:users`, `User.first`
@@ -144,14 +145,18 @@ module Compony
     "#{rails_action_name(...)}_comp"
   end
 
-  # Given a component and a family, this returns the name of the ComponyController action for this component.<br>
+  # Given a component and a family or a component class, this returns the name of the ComponyController action for this component.<br>
   # Optionally can pass a name for extra standalone configs.
   # @param comp_name_or_cst [String,Symbol] Name of the component the action points to.
   # @param model_or_family_name_or_cst [String,Symbol] Name of the family the action points to.
   # @param name [String,Symbol] If referring to an extra standalone entrypoint, specify its name using this param.
   # @see Compony#path
-  def self.rails_action_name(comp_name_or_cst, model_or_family_name_or_cst, name = nil)
-    [name.presence, comp_name_or_cst.to_s.underscore, family_name_for(model_or_family_name_or_cst)].compact.join('_')
+  def self.rails_action_name(comp_name_or_cst_or_class, model_or_family_name_or_cst, name = nil)
+    if comp_name_or_cst_or_class.is_a?(Class) && comp_name_or_cst_or_class <= Compony::Component
+      comp_name_or_cst_or_class = comp_name_or_cst_or_class.comp_name
+      model_or_family_name_or_cst = comp_name_or_cst_or_class.family_name
+    end
+    [name.presence, comp_name_or_cst_or_class.to_s.underscore, family_name_for(model_or_family_name_or_cst)].compact.join('_')
   end
 
   # Given a component and a family/model, this instanciates and returns a button component.